Java files renamed in bulk - follows the naming scheme `FizzBuzz_<FILENAME>`
