+++ Java/FizzBuzz_Scanner.java
