+++ README.md
@@ -74,7 +74,7 @@ Read more about it here: https://www.tomdalling.com/blog/software-design/fizzbuz
 7. [FizzBuzz implementation in HolyC on TempleOS](https://github.com/NLDev/Hacktoberfest-2020-FizzBuzz/blob/master/HolyC/FizzBuzz.HC) by [@Azgeb](https://github.com/Azgeb) - [See PR](https://github.com/NLDev/Hacktoberfest-2020-FizzBuzz/pull/508) for more information!
 8. [Overly complicated FizzBuzz solution (Python)](https://github.com/NLDev/Hacktoberfest-2020-FizzBuzz/blob/master/Python/FizzBuzz-2.py) by [@InitialPosition](https://github.com/InitialPosition)
 9. [Hardcoded FizzBuzz (Python)](https://github.com/NLDev/Hacktoberfest-2020-FizzBuzz/blob/master/Python/FizzBuzz-HardCoded.py) by [@khong612](https://github.com/khong612)
-10. [FizzBuzz in CharArrays, Bytes and Hex Values (Java)](https://github.com/NLDev/Hacktoberfest-2020-FizzBuzz/blob/master/Java/FizzBuzz-Main.java) by [@nimbl0](https://github.com/nimbl0)
+10. [FizzBuzz in CharArrays, Bytes and Hex Values (Java)](/Java/FizzBuzz_CharsBytesHex.java) by [@nimbl0](https://github.com/nimbl0)
 
 <hr>
